move remaining auto-instrumentation tests into subdirectory
